requirements.txt
streamlit
pymongo
passlib
python-dotenv
bleach
emoji